MicroInt applied in fiats

+++ app/views/companion/blockchain/IssueFiat.scala
@@ -16,6 +16,6 @@ object IssueFiat {
     )(Data.apply)(Data.unapply)
   )
 
-  case class Data(from: String, to: String, transactionID: String, transactionAmount: Int, gas: Int, mode: String, password: String)
+  case class Data(from: String, to: String, transactionID: String, transactionAmount: Double, gas: Int, mode: String, password: String)
 
 }
