+++ app/views/companion/master/RedeemFiat.scala
@@ -12,6 +12,6 @@ object RedeemFiat {
     )(Data.apply)(Data.unapply)
   )
 
-  case class Data(redeemAmount: Int, gas: Int, password: String)
+  case class Data(redeemAmount: Double, gas: Int, password: String)
 
 }
